Disassembler/tools: Support for multiple ACPI tables in one file
Matches the support in iASL and AcpiExec.

+++ source/tools/acpixtract/acpixtract.h
@@ -237,7 +237,7 @@ long
 AxProcessOneTextLine (
     FILE                    *OutputFile,
     char                    *ThisSignature,
-    size_t                  ThisTableBytesWritten);
+    unsigned int            ThisTableBytesWritten);
 
 size_t
 AxConvertLine (
